requirements.txt
aiogram==2.25.1
asyncio==3.4.3
fastapi==0.95.1
openai==0.27.6
pydantic==1.10.7
python-telegram-bot==20.2
uvicorn==0.22.0
ptyhon-dotenv==1.0.0